pyproject.toml
[tool.poetry]
name = "openai-ext"
version = "0.0.1"
description = "`openai-ext` is a Meltano utility extension."
authors = ["AJ Steers, Meltano"]
license = "Apache 2.0"
packages = [
    { include = "openai_ext" }
    # uncomment this if you want to include static assets
    # { include = "files_openai-ext.*" }
]
# If you need to static assets with your extension, you can add them here.
# Under a directory path of "files_openai-ext/"
#include = [
#    "files_openai-ext/somefile.thing"
#]

[tool.poetry.dependencies]
python = "<3.12,>=3.7"
PyYAML = "^6.0.0"
click = "^8.1.3"
typer = "^0.6.1"
"meltano.edk"= {git = "https://github.com/meltano/edk.git", rev="main"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
openai_ext = 'openai_ext.main:app'
